Milestone 6 Completed (Dashboard Completed + Backend Setup)

+++ routes/web.php
@@ -302,16 +302,20 @@ Route::middleware(['auth', 'verified'])->group(function () {
                     Route::delete('/roles-destroy/{id?}', 'roleDestroy')->name('roles.destroy');
                     Route::delete('/roles-destroy-by-selection', 'destroyRoleBySelection')->name('roles.destroybyselection');
 
-                    Route::get('/permissions', 'permissionsIndex')->name('permissions.index');
-                    Route::get('/permissions-create', 'permissionCreate')->name('permissions.create');
-                    Route::post('/permissions-store', 'permissionStore')->name('permissions.store');
+                    if (app()->isLocal()) {
+
+                        Route::get('/permissions', 'permissionsIndex')->name('permissions.index');
+                        Route::get('/permissions-create', 'permissionCreate')->name('permissions.create');
+                        Route::post('/permissions-store', 'permissionStore')->name('permissions.store');
+                        Route::get('/permissions-edit/{id?}', 'permissionEdit')->name('permissions.edit');
+                        Route::put('/permissions-update/{id?}', 'permissionUpdate')->name('permissions.update');
+                        Route::delete('/permissions-destroy/{id?}', 'permissionDestroy')->name('permissions.destroy');
+                        Route::delete('/permissions-destroy-by-selection', 'destroyPermissionBySelection')->name('permissions.destroybyselection');
+
+                    }
+
                     Route::get('/permissions-manage/{id?}', 'permissionManage')->name('permissions.manage');
                     Route::put('/permissions-sync/{id?}', 'permissionSync')->name('permissions.sync');
-                    Route::get('/permissions-edit/{id?}', 'permissionEdit')->name('permissions.edit');
-                    Route::put('/permissions-update/{id?}', 'permissionUpdate')->name('permissions.update');
-                    Route::delete('/permissions-destroy/{id?}', 'permissionDestroy')->name('permissions.destroy');
-                    Route::delete('/permissions-destroy-by-selection', 'destroyPermissionBySelection')->name('permissions.destroybyselection');
-
                     // Color Routes
                     Route::get('/colors', 'colorIndex')->name('colors.index');
                     Route::get('/colors-create', 'colorCreate')->name('colors.create');
